src/requirements.txt
amqp==2.1.1
billiard==3.5.0.2
cairocffi==0.7.2
CairoSVG==1.0.20
celery==4.0.0
certifi==2016.8.8
cffi==1.6.0
chardet==2.2.1
cssselect==0.9.1
Cython==0.23.4
dataset==0.4.0
decorator==3.4.0
Django==1.9
django-admin-bootstrapped==2.5.7
django-authtools==1.5.0
django-bootstrap-form==3.1
django-braces==1.8.1
django-classy-tags==0.4
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
django-easy-pdf==0.1.0
django-el-pagination==3.0.1
django-embed-video==1.1.0
django-environ==0.4.0
django-object-tools==1.9
django-openid-auth==0.5
django-paypal==0.3.2
django-registration==0.8
django-social-auth==0.7.23
django-xmlrpc==0.1.5
djangorestframework==3.5.3
easy-thumbnails==2.3
hiredis==0.2.0
html5lib==1.0b10
http-parser==0.8.1
httplib2==0.9.2
IMAPClient==0.10.2
importlib==1.0.3
jsonschema==2.5.1
kombu==4.0.0
lxml==3.6.0
numpy==1.8.1
oauth==1.0.1
oauth2==1.9.0.post1
oauthlib==1.1.2
openpyxl==1.6.2
Orange==2.7.8
Orange-Text==1.2a1
ordereddict==1.1
oursql==0.9.3.1
packaging==16.8
PIL==1.1.7
Pillow==3.4.2
Pinax==0.9a2
Pint==0.7.2
pisa==3.0.33
Pivy==0.5.0
plink==1.9.1
plotly==1.9.3
pp==1.6.4
praw==3.1.0
prettytable==0.6.1
prompt-toolkit==1.0.5
protobuf==3.0.0b2
psutil==1.2.1
psycopg2==2.5
ptyprocess==0.5.1
pudb==2013.3.2
py-bcrypt==0.4
py2neo==3
PyAMF==0.6.1
pyasn1==0.1.7
PyChart==1.39
pycparser==2.14
pycrypto==2.6
pycsp==0.9.0
pycurl==7.19.0
pyDatalog==0.13.0
pydot==1.0.28
pydub==0.9.5
pyenchant==1.6.6
pyfits==3.1.2
pyflakes==0.7.2
pygal==1.2.1
pygeoip==0.3.2
Pygments==2.0.2
pygraphviz==1.1
pyhdf==0.8.3
PyJWT==1.4.1
Pykka==1.1.0
pymc==2.3.2
pymongo==3.2.2
pyodbc==3.0.10
pyOpenSSL==0.13
pyparsing==2.2.0
PyPDF2==1.26.0
pypdfocr==0.7.4
pypdftk==0.3
Pyphen==0.9.4
pypng==0.0.18
pyproj==1.9.3
pyquery==1.2.4
pyramid==1.4.2
PySAL==1.5.0
pyserial==2.6
pyspotify==1.11
PyStemmer==1.3.0
pytesseract==0.1.5
python-amazon-simple-product-api==1.5.0
python-axolotl==0.1.35
python-axolotl-curve25519==0.1
python-dateutil==2.2
python-gflags==2.0
python-http-client==1.2.3
python-inotify==0.5
python-instagram==0.8.0
python-ldap==2.4.13
python-Levenshtein==0.10.2
python-magic==0.4.3
python-mcrypt==1.1
python-mhash==1.4
python-mimeparse==0.1.4
python-openid==2.2.5
python-ptrace==0.9
python-slugify==0.0.7
python-social-auth==0.2.19
pytz==2016.4
PyVirtualDisplay==0.1.2
pywhatsapp==0.0.7
PyYAML==3.11
pyzmq==13.1.0
qrcode==2.7
qtconsole==4.2.1
Quandl==2.2
quantlib==0.1
QuantLib-Python==1.3
rdflib==4.0.1
redis==2.8.0
remix==2.1.3
reportlab==3.3.0
repoze.lru==0.6
requests==2.10.0
requests-cache==0.4.4
requests-oauthlib==0.6.2
resolver-test==1.0.1
restkit==4.2.1
RestrictedPython==3.6.0
rpy2==2.5.2
ruffus==2.5
s3cmd==1.1.0b3
scikit-learn==0.14.1
scikits.statsmodels==0.3.1
scipy==0.14.0
Scrapy==0.16.5
selenium==2.33.0
sendgrid==2.2.1
sendgrid-django==2.0.0
setproctitle==1.1.7
Shapely==1.3.1
simplegeneric==0.8.1
SimpleHMMER==0.2.3
simplejson==3.3.0
simpy==3.0.5
singledispatch==3.4.0.3
six==1.10.0
tinycss==0.3
urllib3==1.5
vine==1.1.3
virtualenv==15.0.3
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
WeasyPrint==0.24
whitenoise==3.2
Whoosh==2.4.1
widgetsnbextension==1.2.6
WTForms==1.0.4
xhtml2pdf==0.2b1



